+++ app/lib/utils/logger.js
@@ -1,11 +1,1 @@
-function info(message) {
-  // eslint-disable-next-line no-console
-  console.log(message);
-}
-
-function error(message) {
-  // eslint-disable-next-line no-console
-  console.error(message);
-}
-
-module.exports = { info, error };
+module.exports = require('nhsuk-bunyan-logger')('gp-data-etl');