+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.1.3 / TBA
+==================
+- Call drain for an empty queue
+- Use bunyan logger
+
 0.1.2 / 2017-10-24
 ==================
 - Add CHANGELOG